requirements.txt
alembic==1.7.5
arrow==0.14.7
beautifulsoup4==4.10.0
black==21.12b0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
css-inline==0.7.6
extruct==0.13.0
fillpdf==0.6.0
Flask==2.0.2
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
greenlet==1.1.2
gunicorn==20.1.0
html-text==0.5.2
html2text==2020.1.16
html5lib==1.1
ics==0.7
idna==3.3
isodate==0.6.1
itsdangerous==2.0.1
Jinja2==3.0.3
jstyleson==0.0.2
lxml==4.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mf2py==1.1.2
mypy-extensions==0.4.3
pathspec==0.9.0
pdf2image==1.16.0
pdfkit==1.0.0
pdfrw==0.4
pdfrw2==0.5.0
Pillow==9.0.0
platformdirs==2.4.1
PyMySQL==1.0.2
pyparsing==3.0.6
pyRdfa3==3.5.3
python-dateutil==2.8.2
python-dotenv==0.19.2
qrcode==7.3.1
rdflib==6.1.1
rdflib-jsonld==0.6.2
recipe-scrapers==13.10.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
TatSu==5.6.1
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
w3lib==1.22.0
webencodings==0.5.1
Werkzeug==2.0.2
wifi-qrcode-generator==0.1
